requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-dotenv==1.1.0
google-generativeai==0.8.5
scikit-learn==1.6.1
aiohttp==3.12.2